requirements.txt
numpy
pandas
openpyxl
flask
flask-cors
sentence-transformers
faiss-cpu
openai
google-generativeai